requirements.txt
Flask 
flask_cors
bcrypt 
python-dotenv
bcrypt
pymongo
